pr review changes for MotifDoc and folder refactoring reverted
