+++ .storybook/MotifDoc/MotifDocContainer.tsx
@@ -1,7 +1,7 @@
 import { DocsContainer } from "@storybook/addon-docs/blocks";
 import "../../src/lib/styles/themes/default-theme.css";
 import { PropsWithChildren } from "react";
-import "./MotifDocStyles.scss";
+import "./styles.scss";
 import { MotifDoc } from "./MotifDoc";
 import { MotifDocContainerProps } from "./types";
